+++ genome/pyrule/__init__.py
@@ -2,13 +2,18 @@
 """
  * @Date: 2023-07-22 15:34:50
  * @LastEditors: Hwrn hwrn.aou@sjtu.edu.cn
- * @LastEditTime: 2023-12-20 21:05:20
+ * @LastEditTime: 2023-12-28 14:05:38
  * @FilePath: /genome/genome/pyrule/__init__.py
  * @Description:
 """
 
 from pathlib import Path
 from typing import Any, Callable
+
+try:
+    from snakemake import main as smk
+except ImportError:
+    from snakemake.cli import main as smk
 
 import snakemake.workflow as _wf