+++ changelog.md
@@ -1,7 +1,7 @@
 <!--
  * @Date: 2023-08-07 15:18:41
  * @LastEditors: Hwrn hwrn.aou@sjtu.edu.cn
- * @LastEditTime: 2023-11-25 23:03:02
+ * @LastEditTime: 2023-11-25 23:04:33
  * @FilePath: /genome/changelog.md
  * @Description:
 -->
@@ -14,6 +14,7 @@ changelog for genome
     - fix:
         - fix metabat2 min contig length
         - fix gff to faa/fa start
+        - fix mantis marker
 - 0.1.4:
     - remove `genome.pyrule.gene`.
     - feat